feat(custom): add 3d configurator for custom-table meja panjang and minor changes
