--- resources/views/components/layouts/xthreed.blade.php
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-        <title>{{ $title ?? 'UD Laris' }}</title>
-
-        <!-- Scripts -->
-        @vite(['resources/css/app.css', 'resources/js/app.js'])
-        <script type='text/javascript' src='http://www.x3dom.org/download/x3dom.js'> </script>
-
-        <!-- Styles -->
-        @livewireStyles
-        <link rel='stylesheet' type='text/css' href='http://www.x3dom.org/download/x3dom.css'></link>
-    </head>
-    <body>
-        @livewire('partials.navbar')
-        <main>
-            {{ $slot }}
-        </main>
-        @livewire('partials.footer')
-        @livewireScripts
-    </body>
-</html>
