--- resources/views/livewire/custom-cabinet-page.blade.php
@@ -1,3 +1,0 @@
-<div>
-    {{-- The whole world belongs to you. --}}
-</div>
