+++ resources/views/livewire/custom-table-page.blade.php
@@ -6,7 +6,7 @@
             <li class="shrink basis-0 flex-1 group">
                 <div class="min-w-7 min-h-7 w-full inline-flex items-center text-xs align-middle">
                     <span
-                        class="size-7 flex justify-center items-center flex-shrink-0 bg-blue-600 font-medium text-white rounded-full dark:bg-gray-700 dark:text-white">
+                        class="size-7 flex justify-center items-center flex-shrink-0 bg-gray-100 font-medium text-gray-800 rounded-full dark:bg-gray-700 dark:text-white">
                         1
                     </span>
                     <div class="ms-2 w-full h-px flex-1 bg-gray-200 group-last:hidden dark:bg-gray-700"></div>
@@ -23,7 +23,7 @@
             <li class="shrink basis-0 flex-1 group">
                 <div class="min-w-7 min-h-7 w-full inline-flex items-center text-xs align-middle">
                     <span
-                        class="size-7 flex justify-center items-center flex-shrink-0 bg-gray-100 font-medium text-gray-800 rounded-full dark:bg-gray-700 dark:text-white">
+                        class="size-7 flex justify-center items-center flex-shrink-0 bg-blue-600 font-medium text-white rounded-full dark:bg-gray-700 dark:text-white">
                         2
                     </span>
                     <div class="ms-2 w-full h-px flex-1 bg-gray-200 group-last:hidden dark:bg-gray-700"></div>
@@ -59,18 +59,18 @@
             <div class="max-w-[85rem] px-4 py-10 sm:px-6 lg:px-8 lg:py-14 mx-auto">
                 <div class="grid sm:grid-cols-1 md:grid-cols-2 lg:grid-cols-2 gap-4 sm:gap-6">
 
-                    @foreach ($categories as $category)
+                    @foreach ($x3d_tables as $x3d_table)
                         <a class="group flex flex-col bg-white border shadow-sm rounded-xl hover:shadow-md transition dark:bg-slate-900 dark:border-gray-800 dark:focus:outline-none dark:focus:ring-1 dark:focus:ring-gray-600"
-                            href="/product-custom/{{ $category->slug }}">
+                            href="/product-custom/meja/{{ $x3d_table->slug }}">
                             <div class="p-4 md:p-5">
                                 <div class="flex justify-between items-center">
                                     <div class="flex items-center">
-                                        <img class="h-[5rem] w-[5rem]" src="{{ url('storage', $category->thumbnail) }}"
-                                            alt="{{ $category->name }}">
+                                        <img class="h-[5rem] w-[5rem]" src="{{ url('storage', $x3d_table->thumbnail) }}"
+                                            alt="{{ $x3d_table->name }}">
                                         <div class="ms-3">
                                             <h3
                                                 class="group-hover:text-blue-600 text-2xl font-semibold text-gray-800 dark:group-hover:text-gray-400 dark:text-gray-200">
-                                                {{ $category->name }}
+                                                {{ $x3d_table->name }}
                                             </h3>
                                         </div>
                                     </div>
